requirements.txt
click==8.1.3
pycryptodome==3.14.1
typer==0.4.1
